Moves SideEffects example, and fills out InlineFunctions class
